test: what the store promises, and what the routes announce
ConfigStore gets a file of its own, out of ConfigTests: update clamps
before saving and before telling anyone, and a listener that writes back
sees the first write and hears each write once. Nothing covered the clamp
in the store before; the Rectangle import test clamped by hand.

The two Router notify tests replaced the harness's didMutate with a
counter, which proved only that the store fires. They watch the change bus
now, the thing the UI actually listens to.
